requirements.txt
flask
sqlalchemy
flask_sqlalchemy
requests
openai
pillow